Rename package from unit to units
