+++ src/units/map.go
@@ -1,4 +1,4 @@
-package unit
+package units
 
 const (
 	Length      = "length"
