+++ src/units/speed.go
@@ -1,4 +1,4 @@
-package unit
+package units
 
 import "github.com/martinlindhe/unit"
